Add files via upload
This is an LSTM based machine learning model, that is is used for energy consumption prediction of an particular region. The data-set used here is taken from kaggle known with the name "AEP_hourly" that has 2 attributes in total, that is current  time and power consumed. The jupyter file has been uploaded for same.